backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.26.0
Pillow>=10.1.0
scikit-learn>=1.4.0
joblib>=1.3.0
